camera2 vndk: Fix onCaptureBufferLost extra / incorrect callbacks.
Before this change, we were not matching native handles against the
CaptureRequest for which a buffer error occurred, and therefore
producing extra and incorrect onCaptureBufferLost callbacks.

Also, the buffer handles returned in the callbacks must be the same as
what the app configured capture sessions with. This way the app can
match the native handle with outputs configured without needing to
inspect the internal contents of the native handle.

Bug: 224810366

Test: Artificially produce frame drops by putting cpus in power saving
      mode; use camera2 vndk; confirm buffer native handles are not
      repeated per frame number.

Change-Id: I4b54e007576d29529170bc192308d01f18b95a13
Signed-off-by: Jayant Chowdhary <jchowdhary@google.com>